v492.0 dev2 - script improvements

+++ scripts/preprocessing/prepro_check_and_clean_files.sh
@@ -24,29 +24,36 @@ move_list()
     done
 }
 
-echo "Error data $FTYPE"
+echo "Searching for errors for data type $FTYPE"
 
 # find all runs with errors and move them
 FLIST=$(grep -irl "error" $FTYPE/*.log)
 if [[ -n $FLIST ]]; then
     file_count=$(echo "$FLIST" | wc -w)
-    echo "FOUND $file_count files with errors"
+    if [[ ! -z $file_count ]]; then
+        echo "FOUND $file_count files with errors"
+    fi
     move_list error "$FLIST"
 fi
 # find all runs with segmentation faults
 FLIST=$(grep -rl "segmentation" $FTYPE/*.log)
 if [[ -n $FLIST ]]; then
     file_count=$(echo "$FLIST" | wc -w)
-    echo "FOUND $file_count files with segmentation faults"
+    if [[ ! -z $file_count ]]; then
+        echo "FOUND $file_count files with segmentation faults"
+    fi
     move_list error "$FLIST"
 fi
 # find all runs without errors and remove them from error directory
 FLIST=$(grep -iL "error" $FTYPE/*.log)
 if [[ -n $FLIST ]]; then
-    echo "FOUND $file_count files without errors"
-    for F in $FLIST; do
-        rm -f ${FTYPE}/error/$(basename $F .log).*
-    done
+    file_count=$(echo "$FLIST" | wc -w)
+    if [[ ! -z $file_count ]]; then
+        echo "FOUND $file_count files without errors - cleaning error directory"
+        for F in $FLIST; do
+            rm -f ${FTYPE}/error/$(basename $F .log).*
+        done
+    fi
 fi
 
 echo "Aux data (and NOTFOUND)"
@@ -55,6 +62,6 @@ if [[ $NAUX -gt 0 ]]; then
     mkdir -p "$FTYPE"/aux
     mv -f "$FTYPE"/*.NOTFOUND "$FTYPE"/aux/
 fi
-echo "Remove lists"
+echo "Remove list file (*.list, *.runlist)"
 rm -f "$FTYPE"/*.list
 rm -f "$FTYPE"/*.runlist
